# [AIRFLOW-XXXX] Slightly speed up mypy runs.