@ashb 2020-02-14
I'm a little bit surprised by the difference, but it was consistent on my laptop. Would appreciate others testing too.