@ashb 2020-02-14
Seems to give a modest speed up on Travis (25s to 19s)