@potiuk 2020-02-14
And BTW  this is totally expect with our codebase if you think of it. It is yet another consequence of our airflow.__init__.py which immediately loads probably 90% of all airflow core files/models etc. On e you modify single file that uses import aitfow.<whatever> you parse and load everything.  I think we should really seriously try to consider changing it. It"s the same reasons n why cli is so slow.