@ashb 2020-02-14
Could be. Would be interesting to see if mypy also chases down "delayed" imports (i.e. ones inside functions) at which point there's only so much we can do. But anyway, mypy time is relatively tiny in comparison to pylint :(