@ashb 2020-02-14
@potiuk Good idea on require_serial -- that has the same effect for all files, but makes a single file "work" and it is still marginally quicker for a single file.