@ashb 2020-02-14
Just for the record: It seems that "dockerising" it adds about an 8x overhead though :(

```
$ time mypy airflow docs/ tests
Success: no issues found in 1831 source files
mypy airflow docs/ tests  0.61s user 0.06s system 99% cpu 0.668 total
```