@potiuk 2020-02-14
How do you compare this? where this 8x comes from? I just see one value.

Is it Linux or mac? 

I believe on mac there is potentially huge docerising impact if we go full "airflow" tests - because then a lot of overhead for just scanning the 10000s of files we have in node_modules (if you have them), for "require_serial" I would expect pretty much 0 overhead from dockerising (once the image is built of course), We have the "build" step that builds all the docker images if needed so there should be no overhead.