@ashb 2020-02-14
That 8x was from comparing `time pre-commit run mypy  -a` (~4.7s) to `mypy airflow docs test` (0.7s)  6.7x difference.

This is on linux, and entirely possibly something about my setup (for instance I'm running docker on btrfs because I'm experimenting.) Anyway. Just an observation, I still think dockerizing these checks by default is the right thing to do.