@potiuk 2020-02-15
Interesting.  I will check it on my Linux :). BTW . Pylint could be - on the other hand - sped up by going parallel - I checked it before and it worked much faster if you had multiple CPUs. However we cannot make it parallel because then introduces randomness in detecting cycles - so again the same story as with mypy - we could get rid of all of the problems by removing stuff from airflow.__init__.py and getting rid of those "pull whole airflow when you import airflow.<whatever>" thing.

I think it is totally unnecessary problem that we created for ourselves by choosing 'airflow' as a public package for all imports from outside (and internally). 

And I think Airflow 2.0 is the right time to change it since we are going to break the compatibility and we can provide automated 'migration assistants' for existing code for 1.0 -> 2.0 migration. 

Maybe we should start discussion on that subject in the devlist? WDYT @ashb?